pyspark_container/pyproject.toml
[tool.poetry]
name = "pyspark_container"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
mastodon-py = "^1.8.1"
six = "^1.16.0"
kafka-python-ng = "^2.2.2"
pyspark = "^3.5.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
